+++ docs/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## v1.31.0
 
+- Settings for TODO - new legs and mythics (rare catches tracker, alert), so far used Ragnarok setting
+- Fishing pet level up prices / + mythic ff
+- Ragnarok HP
+- Killed by ragnarok
+- Share ragnarok loc to ALL CHAT
+- Add new fishing drops to profit tracker, remove old ones
+- Add alerts on ragnarok and Blue Ringed Octopus, Fiery Scuttler, Wiki Tiki
+- Add alerts on new drops?
+- Track catches for Ragnarok in crimson isle tracker, and for 
+- Double hook + short message for all new SC
+- Check treasure messages
+- Count mobs in cap
+- COINS_FISHED_TRIGGERS to work on both new and old versions of SB
+- Ice essense messages?
+- Removed guardian pet from tracker and alerts
+- Glowstone Dust instead of GLowstone
+- Changed colors of good catch and great catch
+- Looking to add a "Hotspot Bait"
+- Autorecombed items support
+
 Released: ???
 
 Features:
@@ -9,6 +29,10 @@ Features:
   - Removed Guardian Pet alert and profit tracker as it's not obtainable from fishing anymore.
   - Removed Shredder and Phantom Rod from Fishing Profit tracker.
   - Added Legendary/Mythic Hermit Crab and Mythic Flying Fish to /feeshPetLevelUpPrices command.
+  - Added new Legendary and Mythic mobs to the alerts and Rare Catches tracker.
+  - Count new mobs towards Sea Creatures Count tracker.
+  - Added new mobs support for compact sea creature catch messages.
+  - Display Ragnarok/s HP in Sea Creatures HP tracker.
 
 
 ## v1.30.0